+++ httpcore/_async/connection.py
@@ -1,5 +1,5 @@
 from ssl import SSLContext
-from typing import List, Optional, Tuple
+from typing import Optional, Tuple, cast
 
 from .._backends.auto import AsyncBackend, AsyncLock, AsyncSocketStream, AutoBackend
 from .._types import URL, Headers, Origin, TimeoutDict
@@ -66,15 +66,18 @@ class AsyncHTTPConnection(AsyncHTTPTransport):
             self._request_lock = self.backend.create_lock()
         return self._request_lock
 
-    async def request(
+    async def arequest(
         self,
         method: bytes,
         url: URL,
         headers: Headers = None,
         stream: AsyncByteStream = None,
-        timeout: TimeoutDict = None,
-    ) -> Tuple[bytes, int, bytes, List[Tuple[bytes, bytes]], AsyncByteStream]:
+        ext: dict = None,
+    ) -> Tuple[int, Headers, AsyncByteStream, dict]:
         assert url_to_origin(url) == self.origin
+        ext = {} if ext is None else ext
+        timeout = cast(TimeoutDict, ext.get("timeout", {}))
+
         async with self.request_lock:
             if self.state == ConnectionState.PENDING:
                 if not self.socket:
@@ -92,9 +95,9 @@ class AsyncHTTPConnection(AsyncHTTPTransport):
 
         assert self.connection is not None
         logger.trace(
-            "connection.request method=%r url=%r headers=%r", method, url, headers
+            "connection.arequest method=%r url=%r headers=%r", method, url, headers
         )
-        return await self.connection.request(method, url, headers, stream, timeout)
+        return await self.connection.arequest(method, url, headers, stream, ext)
 
     async def _open_socket(self, timeout: TimeoutDict = None) -> AsyncSocketStream:
         scheme, hostname, port = self.origin