+++ tests/async_tests/test_interfaces.py
@@ -5,8 +5,13 @@ import pytest
 
 import httpcore
 from httpcore._types import URL
-from tests.conftest import Server, detect_backend
+from tests.conftest import Server
 from tests.utils import lookup_async_backend
+
+
+@pytest.fixture(params=["auto", "anyio"])
+def backend(request):
+    return request.param
 
 
 async def read_body(stream: httpcore.AsyncByteStream) -> bytes:
@@ -20,157 +25,130 @@ async def read_body(stream: httpcore.AsyncByteStream) -> bytes:
 
 
 @pytest.mark.anyio
-async def test_http_request() -> None:
-    async with httpcore.AsyncConnectionPool() as http:
+async def test_http_request(backend: str) -> None:
+    async with httpcore.AsyncConnectionPool(backend=backend) as http:
         method = b"GET"
         url = (b"http", b"example.org", 80, b"/")
         headers = [(b"host", b"example.org")]
-        http_version, status_code, reason, headers, stream = await http.request(
-            method, url, headers
-        )
+        status_code, headers, stream, ext = await http.arequest(method, url, headers)
         await read_body(stream)
 
-        assert http_version == b"HTTP/1.1"
         assert status_code == 200
-        assert reason == b"OK"
+        assert ext == {"http_version": "HTTP/1.1", "reason": "OK"}
         assert len(http._connections[url[:3]]) == 1  # type: ignore
 
 
 @pytest.mark.anyio
-async def test_https_request() -> None:
-    async with httpcore.AsyncConnectionPool() as http:
+async def test_https_request(backend: str) -> None:
+    async with httpcore.AsyncConnectionPool(backend=backend) as http:
         method = b"GET"
         url = (b"https", b"example.org", 443, b"/")
         headers = [(b"host", b"example.org")]
-        http_version, status_code, reason, headers, stream = await http.request(
-            method, url, headers
-        )
+        status_code, headers, stream, ext = await http.arequest(method, url, headers)
         await read_body(stream)
 
-        assert http_version == b"HTTP/1.1"
         assert status_code == 200
-        assert reason == b"OK"
+        assert ext == {"http_version": "HTTP/1.1", "reason": "OK"}
         assert len(http._connections[url[:3]]) == 1  # type: ignore
 
 
 @pytest.mark.anyio
-async def test_request_unsupported_protocol() -> None:
-    async with httpcore.AsyncConnectionPool() as http:
+async def test_request_unsupported_protocol(backend: str) -> None:
+    async with httpcore.AsyncConnectionPool(backend=backend) as http:
         method = b"GET"
         url = (b"ftp", b"example.org", 443, b"/")
         headers = [(b"host", b"example.org")]
         with pytest.raises(httpcore.UnsupportedProtocol):
-            await http.request(method, url, headers)
+            await http.arequest(method, url, headers)
 
 
 @pytest.mark.anyio
-async def test_http2_request() -> None:
-    async with httpcore.AsyncConnectionPool(http2=True) as http:
+async def test_http2_request(backend: str) -> None:
+    async with httpcore.AsyncConnectionPool(backend=backend, http2=True) as http:
         method = b"GET"
         url = (b"https", b"example.org", 443, b"/")
         headers = [(b"host", b"example.org")]
-        http_version, status_code, reason, headers, stream = await http.request(
-            method, url, headers
-        )
+        status_code, headers, stream, ext = await http.arequest(method, url, headers)
         await read_body(stream)
 
-        assert http_version == b"HTTP/2"
         assert status_code == 200
-        assert reason == b"OK"
+        assert ext == {"http_version": "HTTP/2"}
         assert len(http._connections[url[:3]]) == 1  # type: ignore
 
 
 @pytest.mark.anyio
-async def test_closing_http_request() -> None:
-    async with httpcore.AsyncConnectionPool() as http:
+async def test_closing_http_request(backend: str) -> None:
+    async with httpcore.AsyncConnectionPool(backend=backend) as http:
         method = b"GET"
         url = (b"http", b"example.org", 80, b"/")
         headers = [(b"host", b"example.org"), (b"connection", b"close")]
-        http_version, status_code, reason, headers, stream = await http.request(
-            method, url, headers
-        )
+        status_code, headers, stream, ext = await http.arequest(method, url, headers)
         await read_body(stream)
 
-        assert http_version == b"HTTP/1.1"
         assert status_code == 200
-        assert reason == b"OK"
+        assert ext == {"http_version": "HTTP/1.1", "reason": "OK"}
         assert url[:3] not in http._connections  # type: ignore
 
 
 @pytest.mark.anyio
-async def test_http_request_reuse_connection() -> None:
-    async with httpcore.AsyncConnectionPool() as http:
+async def test_http_request_reuse_connection(backend: str) -> None:
+    async with httpcore.AsyncConnectionPool(backend=backend) as http:
         method = b"GET"
         url = (b"http", b"example.org", 80, b"/")
         headers = [(b"host", b"example.org")]
-        http_version, status_code, reason, headers, stream = await http.request(
-            method, url, headers
-        )
+        status_code, headers, stream, ext = await http.arequest(method, url, headers)
         await read_body(stream)
 
-        assert http_version == b"HTTP/1.1"
         assert status_code == 200
-        assert reason == b"OK"
+        assert ext == {"http_version": "HTTP/1.1", "reason": "OK"}
         assert len(http._connections[url[:3]]) == 1  # type: ignore
 
         method = b"GET"
         url = (b"http", b"example.org", 80, b"/")
         headers = [(b"host", b"example.org")]
-        http_version, status_code, reason, headers, stream = await http.request(
-            method, url, headers
-        )
+        status_code, headers, stream, ext = await http.arequest(method, url, headers)
         await read_body(stream)
 
-        assert http_version == b"HTTP/1.1"
         assert status_code == 200
-        assert reason == b"OK"
+        assert ext == {"http_version": "HTTP/1.1", "reason": "OK"}
         assert len(http._connections[url[:3]]) == 1  # type: ignore
 
 
 @pytest.mark.anyio
-async def test_https_request_reuse_connection() -> None:
-    async with httpcore.AsyncConnectionPool() as http:
+async def test_https_request_reuse_connection(backend: str) -> None:
+    async with httpcore.AsyncConnectionPool(backend=backend) as http:
         method = b"GET"
         url = (b"https", b"example.org", 443, b"/")
         headers = [(b"host", b"example.org")]
-        http_version, status_code, reason, headers, stream = await http.request(
-            method, url, headers
-        )
+        status_code, headers, stream, ext = await http.arequest(method, url, headers)
         await read_body(stream)
 
-        assert http_version == b"HTTP/1.1"
         assert status_code == 200
-        assert reason == b"OK"
+        assert ext == {"http_version": "HTTP/1.1", "reason": "OK"}
         assert len(http._connections[url[:3]]) == 1  # type: ignore
 
         method = b"GET"
         url = (b"https", b"example.org", 443, b"/")
         headers = [(b"host", b"example.org")]
-        http_version, status_code, reason, headers, stream = await http.request(
-            method, url, headers
-        )
+        status_code, headers, stream, ext = await http.arequest(method, url, headers)
         await read_body(stream)
 
-        assert http_version == b"HTTP/1.1"
         assert status_code == 200
-        assert reason == b"OK"
+        assert ext == {"http_version": "HTTP/1.1", "reason": "OK"}
         assert len(http._connections[url[:3]]) == 1  # type: ignore
 
 
 @pytest.mark.anyio
-async def test_http_request_cannot_reuse_dropped_connection() -> None:
-    async with httpcore.AsyncConnectionPool() as http:
+async def test_http_request_cannot_reuse_dropped_connection(backend: str) -> None:
+    async with httpcore.AsyncConnectionPool(backend=backend) as http:
         method = b"GET"
         url = (b"http", b"example.org", 80, b"/")
         headers = [(b"host", b"example.org")]
-        http_version, status_code, reason, headers, stream = await http.request(
-            method, url, headers
-        )
+        status_code, headers, stream, ext = await http.arequest(method, url, headers)
         await read_body(stream)
 
-        assert http_version == b"HTTP/1.1"
         assert status_code == 200
-        assert reason == b"OK"
+        assert ext == {"http_version": "HTTP/1.1", "reason": "OK"}
         assert len(http._connections[url[:3]]) == 1  # type: ignore
 
         # Mock the connection as having been dropped.
@@ -180,54 +158,50 @@ async def test_http_request_cannot_reuse_dropped_connection() -> None:
         method = b"GET"
         url = (b"http", b"example.org", 80, b"/")
         headers = [(b"host", b"example.org")]
-        http_version, status_code, reason, headers, stream = await http.request(
-            method, url, headers
-        )
+        status_code, headers, stream, ext = await http.arequest(method, url, headers)
         await read_body(stream)
 
-        assert http_version == b"HTTP/1.1"
         assert status_code == 200
-        assert reason == b"OK"
+        assert ext == {"http_version": "HTTP/1.1", "reason": "OK"}
         assert len(http._connections[url[:3]]) == 1  # type: ignore
 
 
 @pytest.mark.parametrize("proxy_mode", ["DEFAULT", "FORWARD_ONLY", "TUNNEL_ONLY"])
 @pytest.mark.anyio
-async def test_http_proxy(proxy_server: URL, proxy_mode: str) -> None:
+async def test_http_proxy(proxy_server: URL, proxy_mode: str, backend: str) -> None:
     method = b"GET"
     url = (b"http", b"example.org", 80, b"/")
     headers = [(b"host", b"example.org")]
     max_connections = 1
     async with httpcore.AsyncHTTPProxy(
-        proxy_server, proxy_mode=proxy_mode, max_connections=max_connections
+        proxy_server,
+        proxy_mode=proxy_mode,
+        max_connections=max_connections,
+        backend=backend,
     ) as http:
-        http_version, status_code, reason, headers, stream = await http.request(
-            method, url, headers
-        )
+        status_code, headers, stream, ext = await http.arequest(method, url, headers)
         await read_body(stream)
 
-        assert http_version == b"HTTP/1.1"
         assert status_code == 200
-        assert reason == b"OK"
+        assert ext == {"http_version": "HTTP/1.1", "reason": "OK"}
 
 
 @pytest.mark.anyio
-async def test_http_request_local_address() -> None:
-    if lookup_async_backend() == "trio":
+async def test_http_request_local_address(backend: str) -> None:
+    if backend == "auto" and lookup_async_backend() == "trio":
         pytest.skip("The trio backend does not support local_address")
 
-    async with httpcore.AsyncConnectionPool(local_address="0.0.0.0") as http:
+    async with httpcore.AsyncConnectionPool(
+        backend=backend, local_address="0.0.0.0"
+    ) as http:
         method = b"GET"
         url = (b"http", b"example.org", 80, b"/")
         headers = [(b"host", b"example.org")]
-        http_version, status_code, reason, headers, stream = await http.request(
-            method, url, headers
-        )
+        status_code, headers, stream, ext = await http.arequest(method, url, headers)
         await read_body(stream)
 
-        assert http_version == b"HTTP/1.1"
         assert status_code == 200
-        assert reason == b"OK"
+        assert ext == {"http_version": "HTTP/1.1", "reason": "OK"}
         assert len(http._connections[url[:3]]) == 1  # type: ignore
 
 
@@ -249,14 +223,12 @@ async def test_proxy_https_requests(
         max_connections=max_connections,
         http2=http2,
     ) as http:
-        http_version, status_code, reason, headers, stream = await http.request(
-            method, url, headers
-        )
+        status_code, headers, stream, ext = await http.arequest(method, url, headers)
         _ = await read_body(stream)
 
-        assert http_version == (b"HTTP/2" if http2 else b"HTTP/1.1")
         assert status_code == 200
-        assert reason == b"OK"
+        assert ext["http_version"] == "HTTP/2" if http2 else "HTTP/1.1"
+        assert ext.get("reason", "") == "" if http2 else "OK"
 
 
 @pytest.mark.parametrize(
@@ -294,16 +266,17 @@ async def test_connection_pool_get_connection_info(
     keepalive_expiry: float,
     expected_during_active: dict,
     expected_during_idle: dict,
+    backend: str,
 ) -> None:
     async with httpcore.AsyncConnectionPool(
-        http2=http2, keepalive_expiry=keepalive_expiry
+        http2=http2, keepalive_expiry=keepalive_expiry, backend=backend
     ) as http:
         method = b"GET"
         url = (b"https", b"example.org", 443, b"/")
         headers = [(b"host", b"example.org")]
 
-        _, _, _, _, stream_1 = await http.request(method, url, headers)
-        _, _, _, _, stream_2 = await http.request(method, url, headers)
+        _, _, stream_1, _ = await http.arequest(method, url, headers)
+        _, _, stream_2, _ = await http.arequest(method, url, headers)
 
         try:
             stats = await http.get_connection_info()
@@ -324,19 +297,18 @@ async def test_connection_pool_get_connection_info(
     reason="Unix Domain Sockets only exist on Unix",
 )
 @pytest.mark.anyio
-async def test_http_request_unix_domain_socket(uds_server: Server) -> None:
+async def test_http_request_unix_domain_socket(
+    uds_server: Server, backend: str
+) -> None:
     uds = uds_server.config.uds
     assert uds is not None
-    async with httpcore.AsyncConnectionPool(uds=uds) as http:
+    async with httpcore.AsyncConnectionPool(uds=uds, backend=backend) as http:
         method = b"GET"
         url = (b"http", b"localhost", None, b"/")
         headers = [(b"host", b"localhost")]
-        http_version, status_code, reason, headers, stream = await http.request(
-            method, url, headers
-        )
-        assert http_version == b"HTTP/1.1"
+        status_code, headers, stream, ext = await http.arequest(method, url, headers)
         assert status_code == 200
-        assert reason == b"OK"
+        assert ext == {"http_version": "HTTP/1.1", "reason": "OK"}
         body = await read_body(stream)
         assert body == b"Hello, world!"
 
@@ -345,10 +317,10 @@ async def test_http_request_unix_domain_socket(uds_server: Server) -> None:
 @pytest.mark.parametrize("connections_number", [4])
 @pytest.mark.anyio
 async def test_max_keepalive_connections_handled_correctly(
-    max_keepalive: int, connections_number: int
+    max_keepalive: int, connections_number: int, backend: str
 ) -> None:
     async with httpcore.AsyncConnectionPool(
-        max_keepalive_connections=max_keepalive, keepalive_expiry=60
+        max_keepalive_connections=max_keepalive, keepalive_expiry=60, backend=backend
     ) as http:
         method = b"GET"
         url = (b"http", b"example.org", 80, b"/")
@@ -356,7 +328,7 @@ async def test_max_keepalive_connections_handled_correctly(
 
         connections_streams = []
         for _ in range(connections_number):
-            _, _, _, _, stream = await http.request(method, url, headers)
+            _, _, stream, _ = await http.arequest(method, url, headers)
             connections_streams.append(stream)
 
         try:
@@ -371,16 +343,13 @@ async def test_max_keepalive_connections_handled_correctly(
 
 @pytest.mark.anyio
 async def test_explicit_backend_name() -> None:
-    async with httpcore.AsyncConnectionPool(backend=detect_backend()) as http:
+    async with httpcore.AsyncConnectionPool(backend=lookup_async_backend()) as http:
         method = b"GET"
         url = (b"http", b"example.org", 80, b"/")
         headers = [(b"host", b"example.org")]
-        http_version, status_code, reason, headers, stream = await http.request(
-            method, url, headers
-        )
+        status_code, headers, stream, ext = await http.arequest(method, url, headers)
         await read_body(stream)
 
-        assert http_version == b"HTTP/1.1"
         assert status_code == 200
-        assert reason == b"OK"
+        assert ext == {"http_version": "HTTP/1.1", "reason": "OK"}
         assert len(http._connections[url[:3]]) == 1  # type: ignore